New kas::draw::draw_rounded internal module
